backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0
pandas==2.2.2
openpyxl==3.1.4
xlrd==2.0.1
slowapi==0.1.9
markdown==3.6
google-generativeai==0.5.4
groq==0.11.0
sendgrid==6.11.0
python-dotenv==1.0.1
email-validator==2.1.1
